requirements.txt
openai>=1.0.0
anthropic>=0.40.0
python-docx>=1.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
docx2pdf>=0.1.8
